+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>betterbox.mine.game</groupId>
     <artifactId>BetterElo</artifactId>
-    <version>4.8.52-SNAPSHOT</version>
+    <version>4.9.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>BetterElo</name>
